setup ui

+++ data/src/main/java/com/rodrigo/data/dto/PokemonDto.kt
@@ -1,20 +1,20 @@
 package com.rodrigo.data.dto
 
 data class PokemonDto (
-    val id: Long,
+    val id: Int,
     val name: String,
     val displayName: String,
-    val pokemonId: Long,
-    val dexNumber: Long,
-    val generation: Long,
+    val pokemonId: Int,
+    val dexNumber: Int,
+    val generation: Int,
     val formType: String,
     val basePokemon: String,
     val isDefault: Boolean,
     val imageUrl: String,
     val iconUrl: String,
-    val pokeApiId: Long,
+    val pokeApiId: Int,
     val pokeApiUrl: String,
-    val speciesId: Long,
+    val speciesId: Int,
     val speciesUrl: String,
-    val formId: Long
+    val formId: Int
 )
